mongodb-to-gcs/requirements.txt
pymongo
google-cloud-storage